update measure with exported override func
